feature/linux (#273)
* fix darwin.c

* use int

* import foundationnetworking

* drop objc in favor of wrapper class

* fix warning

* wip compiling on linux

* format and fix breaking cast in liinux

* fix test

* ignore tests on linux

* fix linux

* undo process wrapper (wip fixing tests)

* re add wrapper

* fixes

* add import

* linux fixes filehandle

* filehandle fixes

* ignore macos tests

* fix remove side effect

* fixes

* try run install on linux

* fix pipe

* Revert "try run install on linux"

* wip

* fix tests on mac

* fixes

* fixes

* abstract process

* clean up

* wip

* improve foundation import

* improve foundation rewriter

* operator fixes, temp disable code format

* wip

* add region coverage check and more wip for linux

* wip

* use better QOS and add time elapsed to text report

* ignore versioned spm manifest

* update report and fix flake test

* wip fixing remove side effects

* wip

* fix remove side effects

* remove delegate

* remove private import

* lint

* copy dependency

* add which

* improve binary resolution

* update unit tests

* poc project mappings

* update tests

* clean up and fix/add test

* add test, improvements and update acceptance tests

* remove plugin and disable format

* add tests and fix flakiness

* fix linux tests

* delete comment

* lint

* improve package filtering

* delete import

* update readme

* improve token aware visitor replacement

* update dependency

* add crash test case

* document

* update cast

* update cast

* fix

* add video

* fix

* update src

* fix html tag

* try with media attr

* try again

* last chance

* fix

* fix

* undo

* fix tests

* cr comments

+++ Sources/muterCore/TestReporting/MuterTestReport.swift
@@ -24,7 +24,7 @@ struct MuterTestReport {
 
 private extension TimeInterval {
     func formatted() -> String {
-        let time = NSInteger(self)
+        let time = Int(self)
         let ms = Int(truncatingRemainder(dividingBy: 1) * 1000)
         let seconds = time % 60
         let minutes = (time / 60) % 60
